fix: fixes during the merge
- 병합 충돌 해결 및 변수 수정
- 수정 시 부분수정 가능으로 변경
- 본인글만 수정 및 삭제 가능으로 변경
- 시리얼라이저 조정햇서요

+++ articles/views.py
@@ -71,7 +71,14 @@ class ArticleDetailView(APIView):
 
     def put(self, request, pk):
         article = get_object_or_404(Article, pk=pk)
-        serializer = ArticleDetailSerializer(article, data=request.data)
+
+        if article.author != request.user:
+            return Response(
+                data={"detail": "본인의 글만 수정할 수 있습니다!"},
+                status=status.HTTP_403_FORBIDDEN,
+            )
+
+        serializer = ArticleDetailSerializer(article, data=request.data, partial=True)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
@@ -79,6 +86,13 @@ class ArticleDetailView(APIView):
 
     def delete(self, request, pk):
         article = get_object_or_404(Article, pk=pk)
+
+        if article.author != request.user:
+            return Response(
+                data={"detail": "본인의 글만 삭제할 수 있습니다!"},
+                status=status.HTTP_403_FORBIDDEN,
+            )
+
         article.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
